Reuse createImageBitmap polyfill
